Extract Rust HTTP control-plane primitives
